just not imported

+++ download.py
@@ -4,10 +4,10 @@ model_name:str = "runwayml/stable-diffusion-v1-5"
 
 try:
   from diffusers import StableDiffusionPipeline
-  pipeline = StableDiffusionPipeline.from_pretrained(model_name, torch_dtype=torch.float16)
+  pipeline = StableDiffusionPipeline.from_pretrained(model_name)
 except:
   from diffusers import DiffusionPipeline
-  DiffusionPipeline.from_pretrained(model_name,torch_dtype=torch.float16)
+  DiffusionPipeline.from_pretrained(model_name)
 
 
 # pipeline.save_pretrained("./stable-diffusion-v1-5")
